InstructionList is now a List; Factor Instructions out of Path; Add PT Instruction types
